Implement Color class
Was going to use java.awt.Color, but to keep
consistent with book (which lets color component
values act as a sort of HDR by letting the value
range outside 0-1), I'm going with this implementation.

I did add a method to convert to an AWT Color, though.

May refactor to eliminate this class later, if the
"HDR" aspect doesn't seem useful.