Simple app to to demonstrate working shading.
This implements the simple app to render a single sphere
that is outlined in the "Putting it Together" section at
the end of Chapter 6 in the book.

No unit testing is in place for the app, since it will be
replaced by a simple GUI raytracing app after scenes are
implemented in the rtlib.