Convert from Gradle to Maven
Current Gradle only supports Java 16 via toolchains
(i.e., run gradle using Java 15 and then build and
run tests & apps using Java 16 configured as a gradle
toolchain). Gradle 7 (currently on M3, soon to release
RC1) will support Java 16, but a) not sure when it
will be released, and b) I'm not sure I want to mess
will having to always keep N-1 available for gradle
when Java N is released.

OTOH, libgdx (which I've used in the past and might
want to use in the future) has standardized
on Gradle, so I'm checking this in to a topic branch,
and I'll try toolchains and possibly Gradle 7M3 to
see how well they work, then make a final decision
on moving to Maven. 99% certain on it at this point,
though.